# Parallel, multicore processing for TMK