--- a/tmk/.gitignore
+++ b/tmk/.gitignore
@@ -11,10 +11,13 @@ cpp/feat2tmk
 cpp/featdump
 cpp/regtest/
 cpp/tmk-clusterize
+cpp/tmk-clusterize-parallel
 cpp/tmk-compare-two-tmks
 cpp/tmk-hash-video
 cpp/tmk-query
+cpp/tmk-query-parallel
 cpp/tmk-show-pair-offsets
 cpp/tmk-two-level-score
+cpp/tmk-two-level-score-parallel
 cpp/tmkdump
 cpp/vstr2feat