--- a/tmk/cpp/Makefile
+++ b/tmk/cpp/Makefile
@@ -163,7 +163,7 @@ tmk-query: ./bin/tmk-query.cpp $(LIBHDR) libtmk.a
 tmk-query-parallel: ./bin/tmk-query.cpp $(LIBHDR) libtmk.a
 	$(CC) -Xpreprocessor -fopenmp -lomp bin/tmk-query.cpp -o tmk-query-parallel -L. -ltmk
 tmk-query-with-faiss: ./bin/tmk-query-with-faiss.cpp $(LIBHDR) libtmk.a
-	$(CC) -I/usr/local/include/faiss bin/tmk-query-with-faiss.cpp -o tmk-query-with-faiss -L. -L/usr/local/lib -ltmk -lfaiss -Wl,-rpath=/usr/local/lib
+	$(CC) -Xpreprocessor -fopenmp -I/usr/local/include/faiss bin/tmk-query-with-faiss.cpp -o tmk-query-with-faiss -L. -L/usr/local/lib -ltmk -lfaiss -lomp -Wl,-rpath,/usr/local/lib
 tmk-two-level-score: ./bin/tmk-two-level-score.cpp $(LIBHDR) libtmk.a
 	$(CC) bin/tmk-two-level-score.cpp -o tmk-two-level-score -L. -ltmk
 tmk-two-level-score-parallel: ./bin/tmk-two-level-score.cpp $(LIBHDR) libtmk.a