… tmk/cpp/bin/tmk-query.cpp
… tmk/cpp/bin/tmk-query.h
… tmk/cpp/bin/tmk-two-level-score.cpp