Summary
---------

After carefully reviewing the code in `tmk-query-with-faiss.cpp`, I don't think that the current implementation is working as expected.

- FAISS search is limited to the k-nearest neighbors, which is hardcoded to 5. This is very limited for larger datasets.
- TMK comparison is using `computeCosSim`, and I'm not sure the similarity evaluation is similar (or, needs to be similar?) by FAISS
- Inputs for level 1 and level 2 scores are accepted, but not considered or used at all

However, there is a desire to increase performance (e.g., https://github.com/facebook/ThreatExchange/issues/211).

I took a stab at implementing parallel processing with OMP. I left the code base the same, and the OMP library as optional at compile time, for now so this could be merged without disrupting users that may not have OMP installed or if different flags are required on other OS's.

Sample results (3438 TMK files), `c1` = 0.7, `c2` = 0.7:
- `tmk-query-with-faiss` @ 0.000143 sec ## note: an unfair comparison, because c1/c2 are not considered
> note that, total runtime for query-with-faiss vs. query-parallel is around 80%, that is just for the query step
- `tmk-query-parallel` @ 0.323135 sec
- `tmk-query` @ 4.336303 sec

I haven't yet addressed `tmk-find-cluster-reps.cpp`, but the implementation would be nearly identical.

Test Plan
---------

- OMP must be installed. On a Mac, this can be done with `brew install libomp`.
- Run `make parallel` to build the multi-cpu variants. I also updated the faiss compiler directives to work properly on Mac.