@dxdc 2022-03-26
FYI, I have been getting some segfaults with tmk-clusterize-parallel above 3000 tmk files or so. In some earlier iterations, I tried parallelizing only the inner loop and that also gave excellent results (maybe 80-90% as good), but not as good as parallelizing both.

Tweaking the stack size limits seems to resolve it. Completes in 22 seconds for the below query with 3438 tmk files.

```sh
# see:  https://stackoverflow.com/questions/13264274/why-segmentation-fault-is-happening-in-this-openmp-code
ulimit -v 1048576
KMP_STACKSIZE=512m ./tmk-clusterize-parallel --c1 0.9 --c2 0.4 --min 2 ./tmkfiles/*.tmk
```