@dxdc 2022-03-26
Have now also addressed these points:

- FAISS search is limited to the k-nearest neighbors, which is hardcoded to 5. This is very limited for larger datasets.
- Inputs for level 1 and level 2 scores are accepted, but not considered or used at all
- Default fallback to tmk-query if number of points are below the FAISS recommended threshold (see notes in #197). Simple implementation for now.

Regarding:
> TMK comparison is using computeCosSim, and I'm not sure the similarity evaluation is similar (or, needs to be similar?) by FAISS

Would be helpful to get some help on this for CosSim type searches since I'm far from an expert at understanding how FAISS works. I've tried the following, based on some guidance I've found (https://github.com/facebookresearch/faiss/issues/61), which *does* seem to give better slightly better results.

- Use `faiss::IndexFlatIP` instead of `faiss::IndexFlatL2`
- It seems we're supposed to L2-normalize the vectors. I found `faiss::fvec_renorm_L2`, which looks like it could be the correct method? I'm not entirely sure about the parameters. E.g., maybe something like `faiss::fvec_renorm_L2(vector_dimension, num_database_vectors, database.data())`

Still, the current FAISS implementation appears to be trying to make the level 1 search faster, but that is actually the fastest part of the search. The level 2 search is a lot slower, more accurate, and I think is [happening here](https://github.com/facebook/ThreatExchange/blob/f9aff6dd0c90e6c47ffe4151bced4de1676d84f6/tmk/cpp/algo/tmkfv.cpp#L566-L672).