@Dcallies 2022-03-28
Hey Daniel! Thanks a bunch for digging into this! I haven't finished digesting the earlier message you sent me, but as far as I got was concluding that `faiss::fvec_renorm_L2` was the correct method according to the note on cosign distance in the FAISS documentation at https://github.com/facebookresearch/faiss/wiki/MetricType-and-distances#how-can-i-index-vectors-for-cosine-similarity.

Sadly there has been turnover on the project since the original authors have written this module, I think this is going to be make or break by the test plan. I'll ask around internally (@BarrettOlson might be the best reviewer, but we'll all be making it up as it goes), but what would be convincing to me as a test plan is:
1. Show the expected distance with the brute force with X videos (c1 and c2 dists if they matter)
2. A couple of queries with FAISS, show that the correct videos are returned.

Your performance numbers are super useful! This isn't apples-to-apples because it was on PDQ and L1 distances, but here's what a previous author did for benchmarking: #314. You've already put in a lot of work, so as long as we're reasonably certain it's correct, I'm personally willing to merge as-is and will try and get you a fast turnaround.