@Dcallies 2022-03-28
Whoops, no need to run the PDQ indice tests, just was showing how they did the benchmarking. You can see the variables they listed on the top for faiss inputs. 

The thresholds for PDQ would instead be c1 distance for this.