@dxdc 2022-03-28
Ah, OK!

Modified strategy:

- `haystack.txt` now to include arbitrary collection of ~4100 TMK files I have on my machine + all TMK files in `./regtest`
  - additional ~4100 files were only added so as to have sufficient data without aborting with an error.
- `needles.txt` set to `./regtest/chair-19-sd-bar.tmk` (one of the README test cases)
- tmk-query-with-faiss thresholds set to c1=0 c2=1, to return maximal hits

NOTE: `tmk-query-with-faiss` doesn't return all the same movies, but that was true initially. It's only returning the "best" case of a 100% match. As for the comparison of a5863bd, it seems slightly less helpful with this test case:

```
Query  0 ./regtest/chair-19-sd-bar.tmk:
  distance 0.062066 L1 1.000000 L2 1.000000 metadata ./regtest/chair-19-sd-bar.tmk ## original
  distance 0.090200 L1 1.000000 L2 1.000000 metadata ./regtest/chair-19-sd-bar.tmk ## with changes from [a5863bd]
```

Here are the full results. Note that the parallel and non-parallel variants have the same results, but in a different order due to the nature of how it's being run. We don't care about that because ultimately we can just sort the results at the end.

```
Performance comparison for: 4140 TMK files

With FAISS: 'tmk-query-with-faiss'

VECTOR_DIMENSION     256
NUM_DATABASE_VECTORS 4140
NUM_CENTROIDS        257
MINIMUM POINTS NOT MET FOR FAISS: 4140 FOUND vs 10023 EXPECTED
Start training the quantizer:
Training level-1 quantizer
Training level-1 quantizer on 4140 vectors in 256D
WARNING clustering 4140 points to 257 centroids: please provide at least 10023 training points
Training IVF residual
computing residuals
training 4x256 product quantizer on 4140 vectors in 256D
Training PQ slice 0/4
WARNING clustering 4140 points to 256 centroids: please provide at least 9984 training points
Clustering 4140 points in 64D to 256 clusters, redo 1 times, 25 iterations
  Preprocessing in 0.00 s
  Iteration 24 (0.05 s, search 0.05 s): objective=329.363 imbalance=2.207 nsplit=0       
Training PQ slice 1/4
WARNING clustering 4140 points to 256 centroids: please provide at least 9984 training points
Clustering 4140 points in 64D to 256 clusters, redo 1 times, 25 iterations
  Preprocessing in 0.00 s
  Iteration 24 (0.05 s, search 0.05 s): objective=77.461 imbalance=2.638 nsplit=0        
Training PQ slice 2/4
WARNING clustering 4140 points to 256 centroids: please provide at least 9984 training points
Clustering 4140 points in 64D to 256 clusters, redo 1 times, 25 iterations
  Preprocessing in 0.00 s
  Iteration 24 (0.05 s, search 0.04 s): objective=31.4667 imbalance=2.855 nsplit=0       
Training PQ slice 3/4
WARNING clustering 4140 points to 256 centroids: please provide at least 9984 training points
Clustering 4140 points in 64D to 256 clusters, redo 1 times, 25 iterations
  Preprocessing in 0.00 s
  Iteration 24 (0.05 s, search 0.04 s): objective=18.6312 imbalance=2.662 nsplit=0       
IndexIVFPQ::precompute_table: precomputed tables not needed for inner product quantizers
End training the quantizer.
 add_core times: 0.001 17.034 0.455 
imbalance factor: 3.68157
LOAD SECONDS   = 1.420
NEEDLES COUNT  = 1
HAYSTACK COUNT = 4140

CHECK SECONDS = 0.001
Searching for the 15 nearest neighbors

QUERY SECONDS = 0.000193
MEAN QUERY SECONDS = 0.000193
Query results (vector IDs, then distances):
(Note that the nearest neighbor is not always at distance 0 due to quantization errors.)
Query  0 ./regtest/chair-19-sd-bar.tmk:
  distance 0.090200 L1 1.000000 L2 1.000000 metadata ./regtest/chair-19-sd-bar.tmk


 -- Finished: 2.107 seconds

Without FAISS: 'tmk-query-parallel'
LOAD SECONDS   = 1.140
NEEDLES COUNT  = 1
HAYSTACK COUNT = 4140

CHECK SECONDS = 0.002

QUERY FOR ./regtest/chair-19-sd-bar.tmk
  0.727866 0.733801 ./regtest/chair-orig-22-fhd-no-bar.tmk
  1.000000 1.000000 ./regtest/chair-19-sd-bar.tmk
  0.952330 0.961902 ./regtest/chair-orig-22-sd-bar.tmk
  0.895393 0.901900 ./regtest/chair-22-with-small-logo-bar.tmk
  0.727833 0.733890 ./regtest/chair-orig-22-hd-no-bar.tmk
  0.953348 0.962158 ./regtest/chair-22-sd-grey-bar.tmk
  0.989083 0.991138 ./regtest/chair-20-sd-bar.tmk
  0.952936 0.962041 ./regtest/chair-22-sd-sepia-bar.tmk

QUERY SECONDS = 0.010351
MEAN QUERY SECONDS = 0.010351

 -- Finished: 1.793 seconds

Without FAISS: 'tmk-query'
LOAD SECONDS   = 1.147
NEEDLES COUNT  = 1
HAYSTACK COUNT = 4140

CHECK SECONDS = 0.002

QUERY FOR ./regtest/chair-19-sd-bar.tmk
  1.000000 1.000000 ./regtest/chair-19-sd-bar.tmk
  0.989083 0.991138 ./regtest/chair-20-sd-bar.tmk
  0.953348 0.962158 ./regtest/chair-22-sd-grey-bar.tmk
  0.952936 0.962041 ./regtest/chair-22-sd-sepia-bar.tmk
  0.895393 0.901900 ./regtest/chair-22-with-small-logo-bar.tmk
  0.727866 0.733801 ./regtest/chair-orig-22-fhd-no-bar.tmk
  0.727833 0.733890 ./regtest/chair-orig-22-hd-no-bar.tmk
  0.952330 0.961902 ./regtest/chair-orig-22-sd-bar.tmk

QUERY SECONDS = 0.059822
MEAN QUERY SECONDS = 0.059822

 -- Finished: 1.760 seconds
 ```